[android] fix balance number type

+++ android/app/src/main/kotlin/com/olb/centrio/models/User.kt
@@ -16,5 +16,5 @@ class User {
     @SerializedName("image")
     var picture: String? = null
 
-    var balance: Long = 0
+    var balance: Double = 0.0
 }
